+++ packages/brick-kit/src/internal/bindListeners.ts
@@ -12,6 +12,7 @@ import {
   RuntimeBrickElement,
   StoryboardContextItem,
   UseProviderEventHandler,
+  ProviderPollOptions,
 } from "@next-core/brick-types";
 import { handleHttpError, httpErrorToString } from "../handleHttpError";
 import { computeRealValue, setProperties } from "./setProperties";
@@ -21,7 +22,6 @@ import {
   _internalApiGetProviderBrick,
   symbolForParentTemplate,
   RuntimeBrickElementWithTplSymbols,
-  _internalApiGetMicroAppApiOrchestrationMap,
   symbolForParentRefForUseBrickInPortal,
   RuntimeBrick,
 } from "../core/exports";
@@ -30,9 +30,10 @@ import { looseCheckIf, IfContainer } from "../checkIf";
 import { getUrlByAliasFactory } from "./alias";
 import { getMessageDispatcher } from "../core/MessageDispatcher";
 import { PluginWebSocketMessageTopic } from "../websocket/interfaces";
-import { isCustomApiProvider, getArgsOfCustomApi } from "../core/CustomApis";
 import { applyTheme, applyMode } from "../themeAndMode";
 import { clearMenuTitleCache, clearMenuCache } from "./menu";
+import { PollableCallback, PollableCallbackFunction, startPoll } from "./poll";
+import { getArgsOfCustomApi } from "../core/FlowApi";
 
 export function bindListeners(
   brick: HTMLElement,
@@ -646,62 +647,75 @@ async function brickCallback(
     });
     return;
   }
-  let argsFactoryResult = argsFactory(handler.args, context, event, options);
+  let computedArgs = argsFactory(handler.args, context, event, options);
 
-  if ((handler as UseProviderEventHandler).useProvider) {
-    if (isCustomApiProvider((handler as UseProviderEventHandler).useProvider)) {
-      const allMicroAppApiOrchestrationMap =
-        await _internalApiGetMicroAppApiOrchestrationMap();
-      argsFactoryResult = getArgsOfCustomApi(
-        (handler as UseProviderEventHandler).useProvider,
-        allMicroAppApiOrchestrationMap,
-        argsFactoryResult
-      );
-    }
+  if (isUseProviderHandler(handler)) {
+    computedArgs = await getArgsOfCustomApi(handler.useProvider, computedArgs);
   }
-  const task = (): unknown => (target as any)[method](...argsFactoryResult);
-  const { success, error, finally: finallyHook } = handler.callback ?? {};
-  if (success || error || finallyHook) {
-    try {
-      // Try to catch synchronized tasks too.
-      const result = await task();
-      if (success) {
+  const task = (): unknown => (target as any)[method](...computedArgs);
+
+  const {
+    success,
+    error,
+    finally: finallyHook,
+    progress,
+  } = handler.callback ?? {};
+
+  if (!(success || error || finallyHook || progress)) {
+    task();
+    return;
+  }
+
+  const callbackFactory =
+    (
+      eventType: string,
+      specificHandler: BrickEventHandler | BrickEventHandler[]
+    ): PollableCallbackFunction =>
+    (result: unknown) => {
+      if (specificHandler) {
         try {
-          const successEvent = new CustomEvent("callback.success", {
+          const event = new CustomEvent(eventType, {
             detail: result,
           });
-          [].concat(success).forEach((eachSuccess) => {
-            listenerFactory(eachSuccess, context, runtimeBrick)(successEvent);
+          [].concat(specificHandler).forEach((eachHandler) => {
+            listenerFactory(eachHandler, context, runtimeBrick)(event);
           });
         } catch (err) {
-          // Do not throw errors in `callback.success`,
+          // Do not throw errors in `callback.success` or `callback.progress`,
           // to avoid the following triggering of `callback.error`.
           // eslint-disable-next-line
           console.error(err);
         }
-      }
-    } catch (err) {
-      if (error) {
-        const errorEvent = new CustomEvent("callback.error", {
-          detail: err,
-        });
-        [].concat(error).forEach((eachError) => {
-          listenerFactory(eachError, context, runtimeBrick)(errorEvent);
-        });
-      } else {
+      } else if (eventType === "callback.error") {
         // eslint-disable-next-line
-        console.error(err);
+        console.error("Unhandled callback error:", result);
       }
-    } finally {
-      if (finallyHook) {
-        const finallyEvent = new CustomEvent("callback.finally");
-        [].concat(finallyHook).forEach((eachFinally) => {
-          listenerFactory(eachFinally, context, runtimeBrick)(finallyEvent);
-        });
-      }
-    }
+    };
+
+  const pollableCallback: Required<PollableCallback> = {
+    progress: callbackFactory("callback.progress", progress),
+    success: callbackFactory("callback.success", success),
+    error: callbackFactory("callback.error", error),
+    finally: callbackFactory("callback.finally", finallyHook),
+  };
+
+  let poll: ProviderPollOptions;
+  if (isUseProviderHandler(handler)) {
+    poll = computeRealValue(handler.poll, { ...context, event }, true);
+  }
+
+  if (poll?.enabled) {
+    startPoll(task, pollableCallback, poll);
   } else {
-    task();
+    try {
+      // Try to catch synchronized tasks too.
+      const result = await task();
+      pollableCallback.success(result);
+    } catch (err) {
+      pollableCallback.error(err);
+    } finally {
+      pollableCallback["finally"]();
+    }
   }
 }