+++ packages/brick-scripts/src/main.ts
@@ -4,7 +4,6 @@ import path from "path";
 import os from "os";
 import fs from "fs-extra";
 import chalk from "chalk";
-import prettier from "prettier";
 import * as changeCase from "change-case";
 import { getEasyopsConfig } from "@next-core/repo-config";
 import { ask } from "./ask";
@@ -22,19 +21,15 @@ export async function create(flags: AskFlags): Promise<void> {
   let targetType: TargetType;
   let packageName = "";
   let brickName = "";
-  let templateName = "";
   let processorName = "";
   if (flags.provider) {
     targetType = TargetType["A_NEW_PACKAGE_OF_PROVIDERS"];
     packageName = `providers-of-${flags.provider}`;
   } else {
-    ({
-      targetType,
-      packageName,
-      brickName,
-      templateName,
-      processorName,
-    } = await ask(appRoot, flags));
+    ({ targetType, packageName, brickName, processorName } = await ask(
+      appRoot,
+      flags
+    ));
   }
 
   const pkgRoot = path.join(appRoot, targetMap[targetType], packageName);
@@ -50,14 +45,11 @@ export async function create(flags: AskFlags): Promise<void> {
     case TargetType.A_NEW_CUSTOM_TEMPLATE:
       targetRoot = path.join(pkgRoot, "src/custom-templates");
       break;
-    case TargetType.A_NEW_CUSTOM_PROVIDER_BRICK:
+    case TargetType.A_NEW_CUSTOM_PROVIDER:
       targetRoot = path.join(pkgRoot, "src/data-providers");
       break;
     case TargetType.A_NEW_CUSTOM_PROCESSOR:
       targetRoot = path.join(pkgRoot, "src/custom-processors");
-      break;
-    case TargetType.A_NEW_LEGACY_TEMPLATE:
-      targetRoot = path.join(pkgRoot, "src");
       break;
     default:
       targetRoot = path.join(pkgRoot);
@@ -67,7 +59,6 @@ export async function create(flags: AskFlags): Promise<void> {
     targetType,
     packageName,
     brickName,
-    templateName,
     processorName,
     targetRoot,
   });
@@ -101,40 +92,6 @@ export async function create(flags: AskFlags): Promise<void> {
         }
       }
       break;
-    case TargetType.TRANSFORM_A_MICRO_APP:
-      {
-        const microAppPackageJsonPath = path.join(targetRoot, "package.json");
-        const microAppPackageJson = require(microAppPackageJsonPath);
-        Object.assign(microAppPackageJson.scripts, {
-          start:
-            'concurrently -k -n tsc,build "tsc -w --preserveWatchOutput" "node scripts/build.js -w"',
-          prestart: "rimraf dist",
-          prebuild: "rimraf dist && tsc",
-          build: "node scripts/build.js",
-        });
-
-        files.push([
-          microAppPackageJsonPath,
-          JSON.stringify(microAppPackageJson, null, 2),
-        ]);
-
-        const storyboardJsonPath = path.join(targetRoot, "storyboard.json");
-        const storyboardJson = require(storyboardJsonPath);
-        delete storyboardJson.$schema;
-        const srcIndexTs = path.join(pkgRoot, "src", "index.ts");
-        files.push([
-          srcIndexTs,
-          prettier.format(
-            `import { Storyboard } from "@next-core/brick-types";
-
-const storyboard: Storyboard = ${JSON.stringify(storyboardJson)};
-
-export default storyboard;`,
-            { parser: "typescript" }
-          ),
-        ]);
-      }
-      break;
   }
 
   for (const [filePath, content] of files) {
@@ -150,14 +107,14 @@ export default storyboard;`,
   switch (targetType) {
     case TargetType.A_NEW_BRICK:
     case TargetType.A_NEW_CUSTOM_TEMPLATE:
-    case TargetType.A_NEW_CUSTOM_PROVIDER_BRICK:
+    case TargetType.A_NEW_CUSTOM_PROVIDER:
     case TargetType.A_NEW_CUSTOM_PROCESSOR:
     case TargetType.A_NEW_PACKAGE_OF_BRICKS:
       if (brickName || processorName) {
         // 如果是新建构件/自定义provider构件/构件库，需要更新/新建 `index.ts`。
         const srcIndexTs = path.join(pkgRoot, "src/index.ts");
         switch (targetType) {
-          case TargetType.A_NEW_CUSTOM_PROVIDER_BRICK:
+          case TargetType.A_NEW_CUSTOM_PROVIDER:
             fs.appendFileSync(
               srcIndexTs,
               `import "./data-providers/${changeCase.pascalCase(brickName)}";${
@@ -184,7 +141,7 @@ export default storyboard;`,
         if (
           targetType === TargetType.A_NEW_BRICK ||
           targetType === TargetType.A_NEW_CUSTOM_TEMPLATE ||
-          targetType === TargetType.A_NEW_CUSTOM_PROVIDER_BRICK ||
+          targetType === TargetType.A_NEW_CUSTOM_PROVIDER ||
           targetType === TargetType.A_NEW_CUSTOM_PROCESSOR
         ) {
           console.log(
@@ -223,50 +180,6 @@ export default storyboard;`,
         }
       }
       break;
-    case TargetType.A_NEW_LEGACY_TEMPLATE:
-    case TargetType.A_NEW_PACKAGE_OF_LEGACY_TEMPLATES:
-      {
-        // 如果是新建模板/模板库，需要更新/新建 `index.ts`。
-        const srcIndexTs = path.join(pkgRoot, "src", "index.ts");
-        fs.appendFileSync(srcIndexTs, `import "./${templateName}";${os.EOL}`);
-
-        if (targetType === TargetType.A_NEW_LEGACY_TEMPLATE) {
-          console.log(
-            `${chalk.bold("File updated")}: ./${path.relative(
-              process.cwd(),
-              srcIndexTs
-            )}`
-          );
-        }
-      }
-      break;
-    case TargetType.TRANSFORM_A_MICRO_APP:
-      {
-        const storyboardJsonPath = path.join(targetRoot, "storyboard.json");
-        fs.unlinkSync(storyboardJsonPath);
-        console.log(
-          `${chalk.bold.yellow("File removed")}: ./${path.relative(
-            process.cwd(),
-            storyboardJsonPath
-          )}`
-        );
-      }
-      break;
-    case TargetType.I18N_PATCH_A_PACKAGE_OF_LEGACY_TEMPLATES:
-      {
-        const srcIndexTs = path.join(pkgRoot, "src", "index.ts");
-        fs.writeFileSync(
-          srcIndexTs,
-          `import "./i18n";${os.EOL}${fs.readFileSync(srcIndexTs)}`
-        );
-        console.log(
-          `${chalk.bold("File updated")}: ./${path.relative(
-            process.cwd(),
-            srcIndexTs
-          )}`
-        );
-      }
-      break;
     case TargetType.A_NEW_EDITOR_BRICK:
       {
         const editorBricksIndexTs = path.join(
@@ -294,10 +207,7 @@ export default storyboard;`,
     [
       TargetType.A_NEW_PACKAGE_OF_BRICKS,
       TargetType.A_NEW_PACKAGE_OF_LIBS,
-      TargetType.A_NEW_PACKAGE_OF_MICRO_APPS,
       TargetType.A_NEW_PACKAGE_OF_PROVIDERS,
-      TargetType.A_NEW_PACKAGE_OF_LEGACY_TEMPLATES,
-      TargetType.TRANSFORM_A_MICRO_APP,
     ].includes(targetType)
   ) {
     // Run `yarn` after created a new package.