+++ packages/brick-scripts/template/micro-apps-pkg/scripts/generate-version-file.js
@@ -1,3 +1,0 @@
-const { createVersionFile } = require("@next-core/build-config-factory");
-
-createVersionFile();