+++ packages/brick-utils/src/builder/index.ts
@@ -1,2 +1,3 @@
 export * from "./assertions";
 export * from "./normalizeBuilderNode";
+export * from "./normalizeMenu";