+++ packages/babel-runtime-helpers/CHANGELOG.md
@@ -3,13 +3,13 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [1.0.12](https://github.com/easyops-cn/next-core/compare/@next-core/babel-runtime-helpers@1.0.11...@next-core/babel-runtime-helpers@1.0.12) (2021-07-23)
+## [1.0.13](https://github.com/easyops-cn/next-core/compare/@next-core/babel-runtime-helpers@1.0.12...@next-core/babel-runtime-helpers@1.0.13) (2021-08-16)
 
 **Note:** Version bump only for package @next-core/babel-runtime-helpers
 
+## [1.0.12](https://github.com/easyops-cn/next-core/compare/@next-core/babel-runtime-helpers@1.0.11...@next-core/babel-runtime-helpers@1.0.12) (2021-07-23)
 
-
-
+**Note:** Version bump only for package @next-core/babel-runtime-helpers
 
 ## [1.0.11](https://github.com/easyops-cn/next-core/compare/@next-core/babel-runtime-helpers@1.0.10...@next-core/babel-runtime-helpers@1.0.11) (2021-07-02)