+++ packages/jest-config-factory/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.1](https://github.com/easyops-cn/next-core/compare/@next-core/jest-config-factory@0.2.0...@next-core/jest-config-factory@0.2.1) (2021-08-16)
+
+**Note:** Version bump only for package @next-core/jest-config-factory
+
 # [0.2.0](https://github.com/easyops-cn/next-core/compare/@next-core/jest-config-factory@0.1.2...@next-core/jest-config-factory@0.2.0) (2021-08-09)
 
 ### Features